LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_pointerrawupdate_after_adoption_from_child.https-expected.txt

CONSOLE MESSAGE: Viewport argument key "initial-scale:1" not recognized and ignored.


FAIL `pointerrawupdate` events should be fired after the element is moved from the parent doc to a child doc assert_true: expected true got false
